+++ plone/formwidget/masterselect/upgrade.py
@@ -1,4 +1,0 @@
-def v3(context):
-    context.runImportStepFromProfile(
-            'profile-plone.formwidget.masterselect:default',
-            'jsregistry', run_dependencies=False, purge_old=False)